+++ packages/dataviews/CHANGELOG.md
@@ -4,14 +4,21 @@
 
 ### Enhancements
 
+- DataForm: add new details layout. [#72355](https://github.com/WordPress/gutenberg/pull/72355)
+- DatViews list layout: remove link variant from primary actions's button. [#72920](https://github.com/WordPress/gutenberg/pull/72920)
 - DataForm: simplify form normalization. [#72848](https://github.com/WordPress/gutenberg/pull/72848)
 - DataViewsPicker: Add With Modal story. [#72913](https://github.com/WordPress/gutenberg/pull/72913)
 - DataViews Actions: Add the flyout submenu for set-as actions grouping. [#73000](https://github.com/WordPress/gutenberg/pull/73000)
+- DataForm: make the card layout borderless. [#72514](https://github.com/WordPress/gutenberg/pull/72514)
 
 ### Bug fixes
 
 - useFormValidity: make it work with any level of nesting in the form. [#72588](https://github.com/WordPress/gutenberg/pull/72588)
 - Fix: DataViews modal actions in list layout. [#72793](https://github.com/WordPress/gutenberg/pull/72793)
+
+### Enhancements
+
+- DataViews: Group DataViews actions based on primary actions and regular actions and adds a separator between the groups. [#72866](https://github.com/WordPress/gutenberg/pull/72866)
 
 ## 10.2.0 (2025-10-29)